refactor(components): extract methods for later reuse
Refs: #327
